+++ app/Http/Controllers/PictureController.php
@@ -16,11 +16,15 @@ class PictureController extends Controller
 
     public function create(Folder $folder)
     {
+        $this->authorize('admin');
+
         return view("pictures/create", ['folder'=>$folder]);
     }
 
     public function store(Folder $folder, StorePicture $request)
     {
+        $this->authorize('admin');
+
         $request->merge(['folder_id'=>$folder->id]);
         $validated = $request->validated();
         $picture=Picture::create($request->all(['folder_id', 'access', 'link', 'name', 'info', 'alternative', 'slug']));
@@ -29,16 +33,22 @@ class PictureController extends Controller
 
     public function show(Folder $folder, Picture $picture)
     {
+        $this->authorize('show', $picture);
+
         return view('pictures/show', ['folder'=>$folder, 'picture'=>$picture]);
     }
 
     public function edit(Folder $folder, Picture $picture)
     {
+        $this->authorize('admin', $picture);
+
         return view('pictures/edit', ['folder'=>$folder, 'picture'=>$picture]);
     }
 
     public function update(Folder $folder, Picture $picture, StorePicture $request)
     {
+        $this->authorize('admin', $picture);
+
         $validated = $request->validated();
 
         $picture->name = $request->get('name');
@@ -50,6 +60,8 @@ class PictureController extends Controller
 
     public function destroy(Folder $folder, Picture $picture)
     {
+        $this->authorize('admin', $picture);
+
         $picture->delete();
 
         return redirect()-> route('folder.show', [$folder])->with('status', 'La photo a bien été supprimée');
